Added numeric mentions. Username mentions will be automatically pre-parsed into numeric mentions in a future commit. Added #rid for linking to a specific post.
Added a forum_exists function.
Removed the null bytes from the posts
The report system now uses #rid, #tid and numeric mentions instead of raw HTML.
Improved the custom parser to take control characters into account. It also now parses items at the start of the post.
Fixed the padding on the buttons on the profiles.
#tid now takes into account whether the topic it links to exists.